+++ src/traffic.c
@@ -299,7 +299,7 @@ void livetrafficmeter(char iface[32], int mode)
 		if (timespent<=60) {
 			printf("  time             %9"PRIu64" seconds\n", timespent);
 		} else {
-			printf("  time               %7.2f minutes\n", timespent/(float)60);
+			printf("  time               %7.2f minutes\n", timespent/(double)60);
 		}
 
 		printf("\n");
